complete parser testing compatibility and expose fromtext
