+++ python/tvm/relay/__init__.py
@@ -72,5 +72,4 @@ def _debug(*args):
     pdb.set_trace()
 
 # Parser
-parse = parser.parse
-parse_file = parser.parse_file
+fromtext = parser.fromtext
