+++ tests/scripts/task_relay_parser.sh
@@ -7,5 +7,5 @@ rm -rf python/tvm/*.pyc python/tvm/*/*.pyc python/tvm/*/*/*.pyc
 
 make cython || exit -1
 make cython3 || exit -1
-TVM_FFI=cython python -m nose -v tests/python/relay/ir_relay_parser.py || exit -1
-TVM_FFI=ctypes python3 -m nose -v tests/python/relay/ir_relay_parser.py || exit -1
+TVM_FFI=cython python -m nose -v tests/python/relay/test_ir_parser.py || exit -1
+TVM_FFI=ctypes python3 -m nose -v tests/python/relay/test_ir_parser.py || exit -1
